WIP -> workin on overlay manager to get componentview smaller
